+++ libs/draggable/src/hook/index.ts
@@ -1,4 +1,5 @@
+export * from './useComponent';
+export * from './useIcon';
 export * from './useInit';
 export * from './useShare';
 export * from './useShow';
-export * from './useIcon';