+++ libs/map/basemap/vite.config.ts
@@ -26,7 +26,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: '../../../dist/libs/map-basemap',
+    outDir: '../../../dist/libs/map/basemap',
     reportCompressedSize: true,
     commonjsOptions: {
       transformMixedEsModules: true,