+++ libs/share/integrations/vite.config.ts
@@ -26,7 +26,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: '../../../dist/libs/shared-integrations',
+    outDir: '../../../dist/libs/share/integrations',
     emptyOutDir: true,
     reportCompressedSize: true,
     commonjsOptions: {