+++ libs/content-menu/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0 (2025-04-13)
+
+### 🚀 Features
+
+- **content-menu:** add handle mobile ([88cf2b9](https://github.com/hung4564/vue-library/commit/88cf2b9))
+
+### ❤️ Thank You
+
+- hung.pv
+
 ## 0.0.2 (2024-06-17)
 
 ### 🚀 Features