+++ libs/draggable/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.0 (2025-04-13)
+
+### 🚀 Features
+
+- add handle for card container ([1249e2f](https://github.com/hung4564/vue-library/commit/1249e2f))
+
+- **shared:** add new store ([3f6a164](https://github.com/hung4564/vue-library/commit/3f6a164))
+
+### ❤️ Thank You
+
+- hung.pv
+
 ## 0.0.2 (2024-06-17)
 
 This was a version bump only for draggable to align it with other projects, there were no code changes.