+++ libs/share/shared/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.1.0 (2025-04-13)
+
+### 🚀 Features
+
+- list, aside control ([93a607f](https://github.com/hung4564/vue-library/commit/93a607f))
+
+- editor io custom ([d14fca1](https://github.com/hung4564/vue-library/commit/d14fca1))
+
+- **shared:** add new store ([3f6a164](https://github.com/hung4564/vue-library/commit/3f6a164))
+
+### 🩹 Fixes
+
+- doc ([6a97b81](https://github.com/hung4564/vue-library/commit/6a97b81))
+
+### ❤️ Thank You
+
+- hung.pv
+
 ## 0.0.2 (2024-06-17)
 
 This was a version bump only for shared to align it with other projects, there were no code changes.